% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml.load_file.R
\name{yaml.load_file}
\alias{yaml.load_file}
\title{Convert a YAML file into R objects}
\usage{
yaml.load_file(input, error.label, readLines.warn = TRUE, ...)
}
\arguments{
\item{input}{filename or connection; Ff \code{input} is a filename, that file must be encoded in UTF-8.}

\item{error.label}{A label to prepend to error messages (see Details).}

\item{readLines.warn}{logical (default:TRUE) Suppress warnings from readLines used inside read_yaml}

\item{...}{arguments to pass to \code{\link{yaml.load}}}
}
\value{
If the root YAML object is a map, a named list or list with an attribute of 'keys' is
returned.  If the root object is a sequence, a list or vector is returned, depending
on the contents of the sequence.  A vector of length 1 is returned for single objects.
}
\description{
Parse a YAML file and return R objects. Loads file and sends to \code{\link{yaml.load}}.
}
\details{
Helper function for \code{\link{yaml.load}}. Given a file or connections it loads
the contents into memory and calls
\code{\link{yaml.load_file}} which calls \code{\link{yaml.load}} with the
loaded content. See \code{\link{yaml.load}} for full details.
}
\examples{
\dontrun{
# reading from a file (uses readLines internally)
filename <- tempfile()
cat("foo: 123", file=filename, sep="\n")
yaml.load_file(filename)
}
}
\seealso{
Other yaml: 
\code{\link{as.yaml}()},
\code{\link{read_yaml}()},
\code{\link{write_yaml}()},
\code{\link{yaml.load}()}
}
\concept{yaml}
