% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml12.R
\docType{package}
\name{yaml12}
\alias{yaml12}
\title{Methods to Convert R Data to YAML1.2 and Back Again}
\description{
Implements the 'libfyaml' 'YAML' 1.2 parser and emitter
<https://github.com/pantoniou/libfyaml> for R. Majority of
code was originally written by Jeremy Stephens. This is a port
to the libfyaml parser for 1.2 support.
}
\author{
Shawn Garbett \email{shawn.garbett@vumc.org}
}
